pyproject.toml
[project]
name = "games"
version = "0.1.0"
description = "Repo for CIS 6200 course project"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "matplotlib>=3.3.0",
    "pillow>=8.0.0",
    "tqdm>=4.60.0",
    "anthropic",
    "openai",
    "groq",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "pytest"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["src/games"]

[tool.hatch.build.targets.wheel]
packages = ["src/games"]

